# Improve performance of get_model_options