--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -15,6 +15,7 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.1
 Additional_repositories: https://mrc-ide.r-universe.dev 
 Imports:
+    data.table,
     digest,
     docopt,
     geojsonio (>= 0.8.0),
@@ -22,7 +23,7 @@ Imports:
     ids,
     jsonlite (>= 1.2.2),
     lgr,
-    naomi (>= 2.7.18),
+    naomi (>= 2.7.19),
     naomi.options,
     porcelain (>= 0.1.8),
     R6,