--- a/tests/testthat/test-model-options.R
+++ b/tests/testthat/test-model-options.R
@@ -369,8 +369,9 @@ test_that("model options work when survey_mid_calendar_quarter missing", {
 })
 
 test_that("can get survey options & default for different indicators", {
-  survey <- file_object(file.path("testdata", "survey.csv"))
-  prev_options <- get_survey_options(survey, "prevalence")
+  data <- read_csv(file.path("testdata", "survey.csv"))
+  metadata <- naomi::get_metadata()
+  prev_options <- get_survey_options(data, metadata, "prevalence")
   expect_equal(prev_options$default, scalar("DEMO2016PHIA"))
   expect_equal(prev_options$options, list(
     list(
@@ -391,7 +392,7 @@ test_that("can get survey options & default for different indicators", {
     )
   ))
 
-  art_options <- get_survey_options(survey, "art_coverage")
+  art_options <- get_survey_options(data, metadata, "art_coverage")
   expect_equal(art_options$default, scalar("DEMO2016PHIA"))
   expect_equal(art_options$options, list(
     list(id = scalar("DEMO2016PHIA"),
@@ -399,7 +400,8 @@ test_that("can get survey options & default for different indicators", {
   ))
 
   mock_get_indicator_data <- mockery::mock(NULL)
-  recent_infected_options <- get_survey_options(survey, "recent_infected")
+  recent_infected_options <- get_survey_options(data, metadata,
+                                                "recent_infected")
   expect_equal(art_options$default, scalar("DEMO2016PHIA"))
   expect_equal(art_options$options, list(
     list(id = scalar("DEMO2016PHIA"),
@@ -408,9 +410,10 @@ test_that("can get survey options & default for different indicators", {
 })
 
 test_that("getting survey options for missing indicator returns empty values", {
-  survey <- file_object(file.path("testdata", "survey_prevalence_only.csv"))
+  data <- read_csv(file.path("testdata", "survey_prevalence_only.csv"))
+  metadata <- naomi::get_metadata()
   expect_equal(
-    get_survey_options(survey, "art_coverage"),
+    get_survey_options(data, metadata, "art_coverage"),
     list(
       options = NULL,
       default = scalar("")