This will speed up in a few ways
* Reduce duplicate reads (of shape file and geojson)
* Use vectorised version of naomi functions for getting labels (atm this is slow because each call to naomi translates some strings, now that translation is only done once)
* Use data.table to read the csv file, this much faster than read.csv

Running benchmarking locally before these changes on current Cameroon data from ADR we have
```
> microbenchmark::microbenchmark(model_options(input), times = 10)
Unit: seconds
                 expr     min       lq     mean   median       uq      max neval
 model_options(input) 2.14557 2.154384 2.249949 2.232277 2.351745 2.373982    1
```

After this we are getting

```
> microbenchmark::microbenchmark(model_options(input), times = 10)
Unit: milliseconds
                 expr      min       lq    mean   median      uq      max neval
 model_options(input) 376.0353 383.2105 406.114 385.1356 395.946 561.7291    10
```

Looking at the profile the timing is now
* reading geojson `hintr_geojson_read` ~ 120ms
* getting filters from this  `get_region_filters` ~ 50ms
* reading the survey data `read_csv` ~ 40ms
* reading anc data `naomi::read_anc_testing` ~ 20ms
* building the options json `naomi.options::get_controls_json` ~ 80ms

I think this will be a bit quicker live because we have caching of the geojson read.