@jeffeaton 2022-11-17
Comments on this:

* Is `readr::read_csv()` sufficiently performant to use instead of `data.table::fread()`? `readr` is back-ended by `vroom` now, and I think competitive with `data.table`.
* Would it be appropriate to replace some of the reads with the Naomi read functions? E.g. for reading survey data set: https://github.com/mrc-ide/naomi/blob/45f37ef23cd600fe293b07df170c19124248b4d4/R/read-data.R#L40