@r-ash 2022-11-18
> * Is `readr::read_csv()` sufficiently performant to use instead of `data.table::fread()`? `readr` is back-ended by `vroom` now, and I think competitive with `data.table`.

```
> microbenchmark::microbenchmark(readr::read_csv(survey), data.table::fread(survey, nThread = 1), vroom::vroom(survey), times = 10)
```
So for the survey file it is quicker

![Screenshot_20221118_120436](https://user-images.githubusercontent.com/39248272/202701255-234ae912-0e8e-436f-ab0d-df92635f4e37.png)

> * Would it be appropriate to replace some of the reads with the Naomi read functions? E.g. for reading survey data set: https://github.com/mrc-ide/naomi/blob/45f37ef23cd600fe293b07df170c19124248b4d4/R/read-data.R#L40

Ideally yes it would be best if we had a function which reads each file type and we use this everywhere we read, but I'm aware some different behaviour of data reading between naomi and hintr. I thought it would be quite fiddly and slow to make sure that is all consistent and works for both hintr and hint for all input data. Though on quick look it doesn't look like much difference in survey reading at least. I think my preference would be to go with this for now, and we can try tidy up reading later when there is less time pressure